修正:全面程式碼審查後的缺陷修復
圖集快取:
- AtlasHashCalculator 後備鍵移除 GetInstanceID(),快取不再每次啟動失效
- ReleaseAtlasTextures 跳過別名紋理的 Apply(true,true),避免損壞遊戲現役紋理
- RenderTexture 後備路徑的 manifest 尺寸與實際寫出位元組一致
- 載入前驗證原始位元組數(僅攔截截斷),textureNames null 防護

紋理快取:
- IORetryHelper 寫入原子化(tmp + File.Replace)
- 統一 ResizedTextureCache 鎖(新增 GetResizedTextureCacheCopy)
- TryGetCachedTexturePath 磁碟 I/O 移出鎖外
- savedTextures 反向對照表(O(1) 反查),skippedBakingTextureNames 改 ConcurrentDictionary
- PNG 標頭尺寸上限驗證、像素累計改 long、transpiler 零替換時警告

提前載入/相容性:
- GenTypes_AllLeafSubclasses 改 ConcurrentDictionary 並回傳副本
- TranslationInjector 逐檔 try/catch(XmlException 不再中止注入)
- BakingSkipList rootsInitialized 移至迴圈成功後設定
- RedirectHugslib 延遲呼叫包 try/catch、ReloadContentInt 檢查短路

延遲載入/XML:
- 音效佇列確立單一 Unpatch 擁有者,移除過時 TryDequeueSubSound
- 圖形協程改 RestartStopwatch、PostLoadSpecial 僅成功時執行
- GraphicData 子類別跳過快取路徑
- XmlChangeDetector 改順序敏感雜湊折疊(取代可抵消的 Sum)

其他:
- SessionCache.ExposeData null 防護、Startup postfix 尾段例外隔離
- 設定視窗 Resizer null 防護、FGLLog 移除多餘 StackTrace
- 測試:改名無意義測試、RIMWORLD_MANAGED_DIR 環境變數支援

+++ Source/SmarterAtlasBaking/AdaptiveAtlasBaker.cs
@@ -64,7 +64,7 @@ namespace FasterGameLoading
                 var key = kvp.Key;
                 var allTexturesForThisGroup = kvp.Value.Item1.ToList();
 
-                int pixelsInCurrentSlice = 0;
+                long pixelsInCurrentSlice = 0;
                 var batchForNextBake = new List<(Texture2D main, Texture2D mask)>();
                 var bakedAtlasesForGroup = new List<StaticTextureAtlas>();
 
@@ -77,7 +77,8 @@ namespace FasterGameLoading
                         ? m : null;
 
                     batchForNextBake.Add((texture, mask));
-                    pixelsInCurrentSlice += texture.width * texture.height;
+                    // 使用 long 乘積避免大尺寸紋理造成 int 溢位
+                    pixelsInCurrentSlice += (long)texture.width * texture.height;
 
                     if (pixelsInCurrentSlice >= adaptivePixelsPerSlice)
                     {
@@ -147,7 +148,7 @@ namespace FasterGameLoading
             List<(Texture2D main, Texture2D mask)> batch,
             List<StaticTextureAtlas> bakedAtlases,
             Stopwatch bakeStopwatch,
-            int pixelsInThisSlice,
+            long pixelsInThisSlice,
             ref float measuredBakeSpeed,
             ref int adaptivePixelsPerSlice,
             float targetBakeTime,
